refactor(sheets-data-validation): move list option serialization to sheets
