+++ packages/sheets-data-validation/src/facade/f-data-validation-builder.ts
@@ -18,6 +18,7 @@ import type { IDataValidationRule, IDataValidationRuleOptions } from '@univerjs/
 import type { FRange } from '@univerjs/sheets/facade';
 import { DataValidationErrorStyle, DataValidationOperator, DataValidationType, generateRandomId } from '@univerjs/core';
 import { serializeRangeToRefString } from '@univerjs/engine-formula';
+import { serializeListOptions } from '@univerjs/sheets';
 import { FDataValidation } from './f-data-validation';
 
 /**
@@ -791,7 +792,7 @@ export class FDataValidationBuilder {
      */
     requireValueInList(values: string[], multiple?: boolean, showDropdown?: boolean): FDataValidationBuilder {
         this._rule.type = multiple ? DataValidationType.LIST_MULTIPLE : DataValidationType.LIST;
-        this._rule.formula1 = JSON.stringify(values.filter(Boolean));
+        this._rule.formula1 = serializeListOptions(values);
         this._rule.formula2 = undefined;
         this._rule.showDropDown = showDropdown ?? true;
